+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   form answers table
 - Fix - (#6105) Remove WEASIS `Viewer` link in image series resource views
 - Fix - (#6107) Fix link to image series in image series name column
+- Fix - (#6104) Fix migration of versions `item_name` column
 
 ## 7.3.0